separated Send Expression Test in new class
